requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
decorator==5.1.1
distlib==0.3.9
Django==5.1.1
django-extensions==3.2.3
django-jinja==2.11.0
djangorestframework==3.15.2
executing==2.1.0
factory_boy==3.3.1
Faker==29.0.0
filelock==3.16.1
identify==2.6.2
idna==3.10
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nodeenv==1.9.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13